// src/utils/interfaces/IThemeProps.interface.d.ts
export interface IThemeProps {
  background: {
    primary: string;
    secondary: string;
  };
  typography: {
    light: string;
    bold: string;
    bolder: string;
  };
  red: string;
  green: string;
  errorRed: string;
  gray: string;
  disabled: string;
  coolGray: string;
  darkGray: string;
  applaudoNavyblue: string;
  applaudoLightblue: string;
  applaudoLightGray: string;
  applaudoPurple: string;
  applaudoPurple110: string;
  applaudoPurple90: string;
  applaudoGray: string;
  applaudoYellow: string;
  applaudoGreen: string;
  yellow: string;
  lightYellow: string;
  black: string;
  white: string;
  mediumGray: string;
  cyan: string;
  lightGreen: string;
  brown: string;
  brownGray: string;
  lightGray: string;
  redHoverFocused: string;
  lightRed: string;
  redPressed: string;
  grayBorder: string;
  charcoal: string;
  platinum: string;
  success: string;
  warning: string;
  text: {
    primary: string;
    secondary: string;
    darkSecondary: string;
  };
  breakpoints: {
    '3xl': string;
    '2xl': string;
    xl: string;
    lg: string;
    md: string;
    sm: string;
    xs: string;
  };
}